Fix test_disconnect by putting it in a new module
