+++ tests/test_endtoend.py
@@ -8,7 +8,6 @@ from pprint import (
 )
 from time import (
     time,
-    sleep,
 )
 from json import (
     loads,
@@ -25,12 +24,6 @@ from subprocess import (
 
 import pytest
 
-from .parameterize_utils import (
-    INJECT_TCP_METHOD,
-    NEW_DEPLOYMENT_OPERATION,
-    NoTaggedValue,
-    Probe,
-)
 from .conftest import (
     with_probe,
     after_probe,
@@ -498,44 +491,3 @@ def get_pods(ident):
         "-o", "json",
         "--export",
     )
-
-
-def test_disconnect(request):
-    """
-    Telepresence exits with code 3 if its connection to the cluster is lost.
-    """
-    # Avoid using the Probe fixture because it is scoped for multi-test use to
-    # allow a Telepresence session to be used by multiple tests.  This test is
-    # going to ruin its Telepresence session.  We don't want that to affect
-    # other tests.
-    #
-    # Just pick a semi-arbitrary Probe configuration.  We do need to have
-    # kubectl available in the Telepresence execution context for
-    # ``disconnect_telepresence`` to work, though.
-    probe = Probe(request, INJECT_TCP_METHOD, NEW_DEPLOYMENT_OPERATION)
-    request.addfinalizer(probe.cleanup)
-
-    probe_result = probe.result()
-    disconnect_telepresence(
-        probe_result,
-        probe_result.deployment_ident.namespace,
-    )
-    for i in range(30):
-        returncode = probe_result.telepresence.poll()
-        if returncode is not None:
-            break
-        sleep(1.0)
-
-    with pytest.raises(NoTaggedValue):
-        # Read so we get the last bit of Telepresence output logged.  There is
-        # no tagged value from this command so the read will fail but that's
-        # okay.
-        probe_result.read()
-
-    assert returncode == 3, (
-        "Telepresence returncode did not indicate disconnection detected."
-    )
-
-
-def disconnect_telepresence(probe_result, namespace):
-    probe_result.write("disconnect-telepresence " + namespace)
